src/requirements.txt
IPython
Pillow
cv2
mat73
matplotlib
natsort
numpy
psutil
pytz
scipy
skimage
torch
tqdm